Introduce initial workflow scaffold
This is the initial proposal for a global CTM workflow. It is a lower-common-denominator of the documentation we created for the CitySpace project.

This is meant to be a jumping off point for more experienced CoA developers to contribute to, learn from, and refine. Things like server/environment names specified here might not be the names that developers use in real life.

The "template" linked to doesn't exist yet but will in a future feature branch & PR.