Add proposed oversight guidelines
I think this is a good starting point for keeping developers accountable for following best practices.

Ultimately, the whole dev process revolves around the Pull Request, so whether folks are using the command line, a GUI tool, or a meticulously clean workflow, it doesn't really matter what happens before their branch is ready for PR. All that matters is that the PR is clean and that it is formally reviewed. This outline strives to enforce that by making managers the final check in the process without requiring code-level review.