Create README boilerplate
This is the initial attempt at a universally-applicable README that folks can copy into their projects. Some fields may not apply to all projects, so developers are encouraged to use this more as a guideline than a rigid template.

GitHub's `<a>` tag styling doesn't include an underline, so I had to explicitly state the URL as a key/value pair. If these guides end up getting incorporated into a custom CoA UI then it might be nice to move the link to the `<h3>` environment name.