Convert CitySpace README to new boilerplate template
I wanted to demonstrate how a real project would look in this new format, and will solicit feedback from the CitySpace team on whether they think this is a better approach than having a separate `workflow.md` doc. The hope is that workflow is generally universal at the organization level, and that any project-specific nuances can be addressed directly in a relevant section in the README rather than demanding an entire new file.